myapp/requirements.txt
plotly
hydralit
streamlit-authenticator
streamlit_option_menu
firebase_admin